fix: preserve count-by-group grouping and ordering intent
